FIX: more compiler warnings #2

+++ yappi/mem.c
@@ -32,7 +32,7 @@ void YMEMLEAKCHECK(void)
 #endif
 
 
-unsigned long
+size_t
 ymemusage(void)
 {
     return memused;
